requirements.txt
torch>=2.2
torchvision>=0.17
numpy>=1.24
pandas>=2.0
scikit-learn>=1.3
pillow>=10.0
opencv-python>=4.8
rapidfuzz>=3.6
gradio>=4.0